+++ Dockerfile
@@ -10,10 +10,8 @@ RUN go mod download
 # Copy source code
 COPY . .
 
-# Build both binaries
-RUN go build -o /app/bin/process_raid ./cmd/process_raid
-RUN go build -o /app/bin/update_from_schaledb ./cmd/update_from_schaledb
-RUN go build -o /app/bin/total_analysis ./cmd/total_analysis
+# Build single binary
+RUN go build -o /app/bin/batorment .
 
 # Runtime stage
 FROM debian:bookworm-slim
@@ -26,10 +24,8 @@ RUN apt-get update && apt-get install -y --no-install-recommends \
 
 WORKDIR /app
 
-# Copy binaries from builder
-COPY --from=builder /app/bin/process_raid /app/bin/process_raid
-COPY --from=builder /app/bin/update_from_schaledb /app/bin/update_from_schaledb
-COPY --from=builder /app/bin/total_analysis /app/bin/total_analysis
+# Copy binary from builder
+COPY --from=builder /app/bin/batorment /app/bin/batorment
 
 # Copy entrypoint script
 COPY docker-entrypoint.sh /app/docker-entrypoint.sh
