--- internal/types/config_types.go
@@ -1,11 +1,0 @@
-package types
-
-// Configuration for the Postgres database
-type PostgresConfig struct {
-	Host     string
-	Port     int
-	User     string
-	Password string
-	DBName   string
-	SSLMode  string
-}
