+++ internal/ui/logger.go
@@ -1,0 +1,5 @@
+package ui
+
+import "github.com/BeaverHouse/go-common/logger"
+
+var Log = logger.NewSimpleLogger()
