+++ main.go
@@ -1,0 +1,7 @@
+package main
+
+import "ba-torment-data-process/cmd"
+
+func main() {
+	cmd.Execute()
+}
